+++ .gitignore
@@ -3,6 +3,7 @@
 #
 
 .vscode
+.idea
 
 
 # Binaries for programs and plugins
