+++ internal/services/openmeteo/contract.go
@@ -1,0 +1,7 @@
+package openmeteo
+
+import "net/http"
+
+type httpClient interface {
+	Get(string) (*http.Response, error)
+}
