+++ internal/services/openmeteo/mocks/mocks.go
